feat/db

+++ swag/db/models/product.py
@@ -17,7 +17,7 @@ class Product(db.Model):
     id:Mapped[int] = mapped_column(primary_key=True)
     photo_url: Mapped[str] = mapped_column(String(200), nullable=False)
     name: Mapped[str] = mapped_column(String(50), nullable=False)
-    price:Mapped[float] = mapped_column(Float(), nullable=False)
+    price:Mapped[int] = mapped_column(Integer(), nullable=False)
     quantity: Mapped[int] = mapped_column(Integer(), nullable=False)
     description: Mapped[str] = mapped_column(String(200), nullable=False)
     category: Mapped[ProductCategory] = mapped_column(Enum(ProductCategory), nullable=False)
